__experiments__/rasa_bank/requirements.txt
pydantic-settings==2.3.4
pydantic_core==2.18.4
tinydb==4.8.0
Faker==26.0.0
fastapi==0.111.0
python-dotenv==1.0.1
